Serving refactor cypress: Re-add deploy from registry button link (#4623)
* Make and use nav back to registry button

* Fix gap being gone when aria-disabled true
