+++ frontend/packages/modelServing/src/components/projectDetails/NoModelsView.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import EmptyDetailsView from '@odh-dashboard/internal/components/EmptyDetailsView';
 import { ProjectObjectType, typedEmptyImage } from '@odh-dashboard/internal/concepts/design/utils';
 import type { ProjectKind } from '@odh-dashboard/internal/k8sTypes';
+import { NavigateBackToRegistryButton } from '@odh-dashboard/internal/concepts/modelServing/NavigateBackToRegistryButton';
 import { ModelServingPlatform } from '../../concepts/useProjectServingPlatform';
 import { DeployButton } from '../deploy/DeployButton';
 
@@ -16,8 +17,6 @@ export const NoModelsView: React.FC<{ platform: ModelServingPlatform; project: P
     title={platform.properties.deployedModelsView.startHintTitle}
     description={platform.properties.deployedModelsView.startHintDescription}
     createButton={<DeployButton project={project} />}
-    // footerExtraChildren={
-    //   deployingFromRegistry stuff
-    // }
+    footerExtraChildren={<NavigateBackToRegistryButton isEmptyStateAction />}
   />
 );
